CORE-571: Move test work to test-helpers folder
