+++ test-helpers/common-page-rendering.js
